FunctionCalling/pyproject.toml
[project]
name = "task-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=6.6.0",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]
